feat:タンクからの魔素取り出しの時に過剰に魔素を要求していたためにオーバーロードボーナスを一時無効化。 チャンクの魔素濃度再計算ロジック追加。 魔法陣にインベントリまで実装。

+++ src/main/java/com/github/sweetfish111/reincarnated/magic/nodes/sensor/GetMaxMaso.java
@@ -19,7 +19,7 @@ public class GetMaxMaso extends AbstractMagicNode {
         Object rawData = pullData(0, context);
         if(rawData instanceof ServerPlayer player){
             PlayerMagicData magicData = player.getData(ModAttachments.PLAYER_MAGIC_DATA);
-            return magicData.getMaxMaso();
+            return (double)magicData.getMaxMaso();
         }
         return 0;
     }
